Jannsen requires now PHP 7.1
Model: added prepareStatement() as alias of makeBasicSelect() to improve code readability

Request: fix fixpath() function, now it removes the / in url

Response: fixed add() name in FlashMessage calls

Route: remove / in url beginning before encrypt

Ruleset: fixed addRule function to receive the correct string parameter for the mapping of the field in the rule, keeping it optional.

Validator:  added isodate date format for validation in DATE type. Now the validator accepts date in YYYY-MM-DD and YYYYMMDD format. Added static methods validateRuleStatic() and validateTypeStatic() to allow use of validator in any part of app.

JsonResponse: set forcefully application/json header to the output when this class is used to render content.